rename tiled types file
