+++ src/tiled/TiledTypes.ts
