+++ src/tiled/index.ts
@@ -1,3 +1,3 @@
-export * from './types';
+export * from './TiledTypes';
 export {TiledMapper} from './TiledMapper';
 export {TiledSpriteSheet} from './TiledSpriteSheet';
